refactoring api folder arrangement
