+++ go.mod
@@ -3,10 +3,7 @@ module github.com/sdeoras/api
 go 1.12
 
 require (
-	github.com/golang/protobuf v1.3.0
-	golang.org/x/net v0.0.0-20190301231341-16b79f2e4e95
-	golang.org/x/sys v0.0.0-20190306220723-b294cbcfc56d // indirect
-	golang.org/x/text v0.3.1-0.20180807135948-17ff2d5776d2 // indirect
-	google.golang.org/genproto v0.0.0-20190306222511-6e86cb5d2f12 // indirect
-	google.golang.org/grpc v1.19.0
+	github.com/golang/protobuf v1.3.1
+	golang.org/x/net v0.0.0-20190324223953-e3b2ff56ed87
+	google.golang.org/grpc v1.19.1
 )
