+++ pb/api.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package api;
+package pb;
 
 // Lambda service is a provider of various lambda services
 service Lambda {
